+++ src/commands/command/get-item.ts
@@ -1,23 +1,14 @@
-import {
-  GetCommand,
-  type GetCommandInput,
-  type GetCommandOutput,
-} from "@aws-sdk/lib-dynamodb";
+import { GetCommand } from "@aws-sdk/lib-dynamodb";
 import type { Attributes, KeyAttributes } from "../../types.js";
 import { GetItemCodec } from "./codecs/get-item.js";
-import { Command } from "./command.js";
+import { AbstractCommand } from "./command.js";
 import type { GetItemInput } from "./inputs/get-item.js";
 import type { GetItemOutput } from "./outputs/get-item.js";
 
 export class GetItem<
   T extends Attributes = Attributes,
   K extends KeyAttributes = KeyAttributes,
-> extends Command<
-  GetItemInput<K>,
-  GetItemOutput<T>,
-  GetCommandInput,
-  GetCommandOutput
-> {
+> extends AbstractCommand<GetItemInput<K>, GetItemOutput<T>, GetCommand> {
   constructor(input: GetItemInput<K>) {
     super({
       input,