+++ src/commands/command/inputs/index.ts
@@ -1,2 +1,3 @@
+export type { CreateTableInput } from "./create-table.js";
 export type { DeleteItemInput, DeleteItemReturnValue } from "./delete-item.js";
 export type { GetItemInput } from "./get-item.js";