+++ src/commands/command/outputs/create-table.ts
@@ -1,0 +1,7 @@
+import type { CreateTableCommandOutput } from "@aws-sdk/client-dynamodb";
+
+export type CreateTableOutput = Record<string, never>;
+
+export function decode(_: CreateTableCommandOutput): CreateTableOutput {
+  return {};
+}