+++ src/commands/command/outputs/index.ts
@@ -1,2 +1,3 @@
+export type { CreateTableOutput } from "./create-table.js";
 export type { DeleteItemOutput } from "./delete-item.js";
 export type { GetItemOutput } from "./get-item.js";