+++ test/integration/commands/get-item.spec.ts
@@ -12,7 +12,7 @@ describe(DynamoDbClient.name, () => {
       const table = "test-table";
       await client.createTable({
         name: table,
-        primaryKey: {
+        keySchema: {
           partitionKey: { name: "pk", type: "S" },
         },
       });
@@ -37,7 +37,7 @@ describe(DynamoDbClient.name, () => {
       const table = "test-table";
       await client.createTable({
         name: table,
-        primaryKey: {
+        keySchema: {
           partitionKey: { name: "pk", type: "S" },
           sortKey: { name: "sk", type: "N" },
         },