+++ test/integration/commands/write-transaction.spec.ts
@@ -17,7 +17,7 @@ describe(DynamoDbClient.name, () => {
       const table = "test-table";
       await client.createTable({
         name: table,
-        primaryKey: {
+        keySchema: {
           partitionKey: { name: "pk", type: "S" },
         },
       });